+++ src/SprykerCommunity/Zed/WebhookProcessor/Business/WebhookProcessorFacade.php
@@ -67,19 +67,4 @@ class WebhookProcessorFacade extends AbstractFacade implements WebhookProcessorF
             ->sendToQueue($webhookMessageTransfer, $queueName);
     }
 
-    /**
-     * {@inheritDoc}
-     *
-     * @api
-     *
-     * @param int $timeLimit
-     *
-     * @return void
-     */
-    public function runWebhookInboxWorker(int $timeLimit): void
-    {
-        $this->getFactory()
-            ->createWebhookInboxWorker()
-            ->run($timeLimit);
-    }
 }